+++ histoire.config.ts
@@ -3,6 +3,11 @@ import { HstVue } from '@histoire/plugin-vue';
 
 export default defineConfig({
     setupFile: 'histoire.setup.ts',
+    storyIgnored: [
+        '**/dist/**',
+        '**/node_modules/**',
+        '**/src/**/_*'
+    ],
     plugins: [
         HstVue(),
     ],