+++ src/components/_SkeletonStd.story.vue
@@ -1,0 +1,14 @@
+<template>
+    <Story title="Story/Skeleton">
+        <Variant title="Default">
+            Work in Progress
+        </Variant>
+    </Story>
+</template>
+
+<script lang="ts" setup>
+
+</script>
+
+<docs lang="md">
+</docs>