+++ src/components/data-list/DataList.story.vue
@@ -1,10 +1,20 @@
 <template>
     <Story title="Components/DataList/DataList">
-        Work in Progress
+        <Variant title="Default">
+            <DataList>
+                <template #mobile>
+                    Mobile
+                </template>
+                <template #desktop>
+                    Desktop
+                </template>
+            </DataList>
+        </Variant>
     </Story>
 </template>
 
 <script lang="ts" setup>
+import DataList from './DataList.vue';
 
 </script>